+++ MvvmPlugin/MvvmPlugin/src/dotnet/ReSharperPlugin.MvvmPlugin/Models/PluginConstants.cs
@@ -14,7 +14,7 @@ public static class PluginConstants
     public const string PlaceHolderName = "__";
 
     public const string Xaml = "XAML";
-
+   
 
 }